Drop indexed before flushing.  Drop indexes in all cases
