+++ db/indexes_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/couchbase/gocb"
 	"github.com/couchbase/sync_gateway/base"
 	"github.com/couchbaselabs/go.assert"
-	"sync"
 )
 
 func TestInitializeIndexes(t *testing.T) {
@@ -20,7 +19,10 @@ func TestInitializeIndexes(t *testing.T) {
 	defer testBucket.Close()
 	defer tearDownTestDB(t, db)
 
-	dropErr := dropAllBucketIndexes(testBucket)
+	goCbBucket, isGoCBBucket := base.AsGoCBBucket(testBucket)
+	assert.True(t, isGoCBBucket)
+
+	dropErr := base.DropAllBucketIndexes(goCbBucket)
 	assertNoError(t, dropErr, "Error dropping all indexes")
 
 	initErr := InitializeIndexes(testBucket, db.UseXattrs(), 0)
@@ -31,89 +33,90 @@ func TestInitializeIndexes(t *testing.T) {
 
 }
 
-// Reset bucket state
-func dropAllBucketIndexes(bucket base.Bucket) error {
-
-	gocbBucket, ok := base.AsGoCBBucket(bucket)
-	if !ok {
-		return fmt.Errorf("Bucket is not gocb bucket: %T", bucket)
-	}
-
-	// Retrieve all indexes
-	indexes, err := getIndexes(bucket)
-	if err != nil {
-		return err
-	}
-
-	wg := sync.WaitGroup{}
-	wg.Add(len(indexes))
-
-	asyncErrors := make(chan error, len(indexes))
-	defer close(asyncErrors)
-
-	for _, index := range indexes {
-
-		go func(indexToDrop string) {
-
-			defer wg.Done()
-
-			log.Printf("Dropping index %s...", indexToDrop)
-			dropErr := gocbBucket.DropIndex(indexToDrop)
-			if dropErr != nil {
-				asyncErrors <- dropErr
-			}
-			log.Printf("...successfully dropped index %s", indexToDrop)
-		}(index)
-
-	}
-
-	// Wait until all goroutines finish
-	wg.Wait()
-
-	// Check if any errors were put into the asyncErrors channel.  If any, just return the first one
-	select {
-	case asyncError := <-asyncErrors:
-		return asyncError
-	default:
-	}
-
-	return nil
-}
-
-// Get a list of all index names in the bucket
-func getIndexes(bucket base.Bucket) (indexes []string, err error) {
-
-	indexes = []string{}
-
-	gocbBucket, ok := base.AsGoCBBucket(bucket)
-	if !ok {
-		return indexes, fmt.Errorf("Bucket is not gocb bucket: %T", bucket)
-	}
-
-	// Retrieve all indexes
-	getIndexesStatement := fmt.Sprintf("SELECT indexes.name from system:indexes where keyspace_id = %q", gocbBucket.GetName())
-	n1qlQuery := gocb.NewN1qlQuery(getIndexesStatement)
-	results, err := gocbBucket.ExecuteN1qlQuery(n1qlQuery, nil)
-	if err != nil {
-		return indexes, err
-	}
-
-	// Close the results in a defer, and set the value of the "err" return value
-	defer func() {
-		err = results.Close()
-	}()
-
-	var indexRow struct {
-		Name string
-	}
-
-	for results.Next(&indexRow) {
-		indexes = append(indexes, indexRow.Name)
-	}
-
-	return indexes, err
-
-}
+//
+//// Reset bucket state
+//func dropAllBucketIndexes(bucket base.Bucket) error {
+//
+//	gocbBucket, ok := base.AsGoCBBucket(bucket)
+//	if !ok {
+//		return fmt.Errorf("Bucket is not gocb bucket: %T", bucket)
+//	}
+//
+//	// Retrieve all indexes
+//	indexes, err := getIndexes(bucket)
+//	if err != nil {
+//		return err
+//	}
+//
+//	wg := sync.WaitGroup{}
+//	wg.Add(len(indexes))
+//
+//	asyncErrors := make(chan error, len(indexes))
+//	defer close(asyncErrors)
+//
+//	for _, index := range indexes {
+//
+//		go func(indexToDrop string) {
+//
+//			defer wg.Done()
+//
+//			log.Printf("Dropping index %s...", indexToDrop)
+//			dropErr := gocbBucket.DropIndex(indexToDrop)
+//			if dropErr != nil {
+//				asyncErrors <- dropErr
+//			}
+//			log.Printf("...successfully dropped index %s", indexToDrop)
+//		}(index)
+//
+//	}
+//
+//	// Wait until all goroutines finish
+//	wg.Wait()
+//
+//	// Check if any errors were put into the asyncErrors channel.  If any, just return the first one
+//	select {
+//	case asyncError := <-asyncErrors:
+//		return asyncError
+//	default:
+//	}
+//
+//	return nil
+//}
+//
+//// Get a list of all index names in the bucket
+//func getIndexes(bucket base.Bucket) (indexes []string, err error) {
+//
+//	indexes = []string{}
+//
+//	gocbBucket, ok := base.AsGoCBBucket(bucket)
+//	if !ok {
+//		return indexes, fmt.Errorf("Bucket is not gocb bucket: %T", bucket)
+//	}
+//
+//	// Retrieve all indexes
+//	getIndexesStatement := fmt.Sprintf("SELECT indexes.name from system:indexes where keyspace_id = %q", gocbBucket.GetName())
+//	n1qlQuery := gocb.NewN1qlQuery(getIndexesStatement)
+//	results, err := gocbBucket.ExecuteN1qlQuery(n1qlQuery, nil)
+//	if err != nil {
+//		return indexes, err
+//	}
+//
+//	// Close the results in a defer, and set the value of the "err" return value
+//	defer func() {
+//		err = results.Close()
+//	}()
+//
+//	var indexRow struct {
+//		Name string
+//	}
+//
+//	for results.Next(&indexRow) {
+//		indexes = append(indexes, indexRow.Name)
+//	}
+//
+//	return indexes, err
+//
+//}
 
 // Reset bucket state
 func validateAllIndexesOnline(bucket base.Bucket) error {
